requirements.txt
google-generativeai==0.5.4
streamlit==1.35.0
python-dotenv==1.0.1
